test(deps): add PATH-based discovery tests for dependency binaries
Introduce a new test module `path_tests.rs` covering PATH-based dependency-binary discovery helpers including checking for executable presence on PATH with platform-specific considerations (Windows PATHEXT and executable suffixes). Also clean up and move relevant tests from `tests.rs` to keep them better organized. Add support code and mocks for testing executable verification, improving reliability of dependency binary checks on various platforms.
