+++ installer/Cargo.toml
@@ -52,7 +52,7 @@ path = "src/bin/package_dependency_binary.rs"
 ##   testing and are excluded from release builds.
 ## - **API stability**: The test-support surface is not covered by semver
 ##   guarantees and may change between minor releases.
-test-support = []
+test-support = ["dep:temp-env"]
 
 [dependencies]
 camino = { workspace = true }
@@ -66,7 +66,7 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 sha2 = { workspace = true }
 tar = { workspace = true }
-temp-env = { workspace = true }
+temp-env = { workspace = true, optional = true }
 tempfile = { workspace = true }
 thiserror = { workspace = true }
 toml = { workspace = true }
@@ -81,6 +81,7 @@ mockall = { workspace = true }
 rstest = { workspace = true }
 rstest-bdd = { workspace = true }
 rstest-bdd-macros = { workspace = true }
+temp-env = { workspace = true }
 tempfile = { workspace = true }
 whitaker-installer = { path = ".", features = ["test-support"] }
