+++ installer/README.md
@@ -111,12 +111,12 @@ Dependency-tool verification is asymmetric by design:
 - `cargo-dylint` is checked by running `cargo dylint --version`.
 - `dylint-link` is checked by resolving the executable on `PATH` and then
   invoking it with `--help`. The probe injects `RUSTUP_TOOLCHAIN` when the
-  caller has not already set it, which avoids the false negatives from
+  caller has not already set it which avoids the false negatives from
   `dylint-link --version` while still rejecting stale shims and broken scripts.
 
 On Windows, the installer honours `PATHEXT` while scanning `PATH`, so the
 normal Cargo-installed `dylint-link.exe` and other shell-resolved executable
-suffixes are recognised and then verified with the same invocation-based probe.
+suffixes are recognized and then verified with the same invocation-based probe.
 
 The wrappers are:
